requirements.txt
torch
numpy
pillow
torch
torchvision
scikit-learn
fastapi
uvicorn
